Intial work on core primitives for integer arithmetic.
This patch implements shims wrapping __builtin_addc / __builtin_subc
from clang; these operations are defined both for a machine word type (Word)
and for structs wrapping tuples of Words, conforming to a new StaticBuffer
protocol. This is a fair bit of machinery, but seems to produce surprisingly
good codegen so far and allows for considerable code reuse.